+++ src/main/java/faang/school/godbless/calculator/Calculators.java
@@ -2,5 +2,5 @@ package faang.school.godbless.calculator;
 
 @FunctionalInterface
 public interface Calculators {
-    int calculate(int first, int second);
+    long calculate(long first, long second);
 }
